pyproject.toml
[project]
name = "phishing-detector"
version = "1.0.0"
description = "Phishing email detection system with reasoning distillation using GPT-4/Claude and Qwen models"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Security Analytics Team", email = "security@example.com"}
]
license = {text = "MIT"}
keywords = ["phishing", "security", "machine-learning", "reasoning-distillation", "email-security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "transformers==4.55.4",
    "trl==0.22.2",
    "datasets>=3.4.1",
    "accelerate>=0.27.0",
    "bitsandbytes>=0.43.0",
    "peft>=0.9.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "beautifulsoup4>=4.12.0",
    "email-validator>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tensorboard>=2.14.0",
    "wandb>=0.16.0",
    "kafka-python>=2.0.2",
    "flask>=3.0.0",
    "gunicorn>=21.0.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "colorama>=0.4.6",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
gpu = [
    "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git",
]
monitoring = [
    "prometheus-client>=0.17.0",
    "psutil>=5.9.0",
]

[project.urls]
Homepage = "https://github.com/your-org/phishing-detector"
Documentation = "https://phishing-detector.readthedocs.io"
Repository = "https://github.com/your-org/phishing-detector.git"
Issues = "https://github.com/your-org/phishing-detector/issues"

[project.scripts]
phishing-detector = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "config.py"

[tool.hatch.build.targets.wheel]
packages = ["data", "models", "training", "evaluation", "inference", "utils"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | checkpoints
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "gpu: marks tests that require GPU",
]

[tool.coverage.run]
source = ["data", "models", "training", "evaluation", "inference", "utils"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
